requirements.txt
python-dotenv==1.0.1
fastapi==0.115.0
uvicorn==0.30.1
pydantic==2.8.2
httpx==0.27.0
langchain==0.2.14
langchain-core==0.2.32
langchain-community
langchain-groq
langgraph
bs4